test(traces): address #4888 review — run the never-executed seam tests + tighten
Review (proof-reviewer + test + hygiene + security) caught:
- The CH-layer mapper-crossing tests lived in a *.integration.test.ts that
  mocked every boundary yet was routed under the Docker-requiring integration
  config AND excluded from the unit config — so it had executed nowhere.
  Merge those 11 cases into the unit file; they now run and pass (66 -> 77).
- Reserved-key leak check was a stringify-substring match; now key-level.
- Split compound `x!==undefined && byteLength===N` assertions for clear failures.
- buildTraceBlobResolutionDeps error message no longer interpolates tenantId
  (avoids tenant id landing in warn logs via err.message).
- Extract duplicated raw-CH row fixtures to __tests__/fixtures/ch-row-fixtures.ts.
- Fix stale #4215 -> #4888 in TraceService.create JSDoc.

Refs #4888
